cast size_t to unsigned long
git-svn-id: svn://svn.h5l.se/heimdal/trunk/heimdal@14859 ec53bebd-3082-4978-b11e-865c3cabbd6b

+++ appl/ftp/ftp/security.c
@@ -741,7 +741,7 @@ sec_login(char *host)
 
 	tmp = realloc(app_data, (*m)->size);
 	if (tmp == NULL) {
-	    warnx ("realloc %u failed", (*m)->size);
+	    warnx ("realloc %lu failed", (unsigned long)(*m)->size);
 	    return -1;
 	}
 	app_data = tmp;
